English spelling corrections English/strings.po dialog      -> dialogue analog      -> analogue
A dialogue box (GUI - Yes/No/Informational or otherwise) contains dialogue, and as such should be correctly spelt, this project has a en_US strings.po of which this spelling would be correct, but not here).
"dialog" has not become an accepted form in English like "program" has instead of "programme" for use in computing terms.
http://oxforddictionaries.com/definition/english/dialogue

analogue, the in use of opposite of digital, should be spelt -logue at the end.
(the spelling analog is American / en_US).
http://oxforddictionaries.com/definition/english/analogue

Once merges are made with Transifex, en_US can be correctly localised.