[rbp/omxplayer] Add a jpeg re-encode thumbnail pipeline
This PR adds a GPU accelerated pipeline to decode/resize/encode a jpeg without requiring
the decoded pixels to be returned to the arm.

The GPU also decodes and resizes the jpeg in stripes, so can do this work with far less memory.

This significantly reduces ARM CPU required for caching thumbnails,
and significantly reduces the memory required.