[rbp/omxplayer] Allow number minor changes to passthrough audio without closing stream.
With passthough audio, the codec or sample rate changing require closing the audio stream.
The number of channels, bits per sample and bitrate don't affect the physical audio stream and so don't require the stream to be closed.

Also, the WaitCompletion call was incorrect as no EOS has been sent at this point, so it just acts as a (timeout) delay.

With these fixes, number of channels changing is seamless with passthough audio, and the glitch is reduced for non-passthough.