[rbp] Add streamsilence option for hdmi
This adds a (mostly) dummy audio engine for Pi which supports drain allowing the
stream silence gui option to appear.
The OnSettingsChange callback enables or disables the stream silence on the GPU.

Enabling this keeps the receiver awake, and avoids losing the first couple of
seconds of audio at start of playback.
Requires updated firmware.