[rbp/omxplayer] Ignore expected errors when tearing down audio
The previous commit to avoid the error log message for port unpopulated
doesn't actually swallow the error event, so the next command issues will fail.

(currently we don't care as we are cleaning up, but it makes editing this code fragile)

I've also changed the teardown sequence to match the video component which seems safer.