[rbp/omxplayer] Avoid clipping with amplification
The amplification algorithm introduced in #3185 relies on a fixed duration between
submitting samples (where max output values are computed) and samples being output,
so volume can be adjusted in time.

When this is too low, such as at start of stream when buffering is low we can apply
attenuation too late and so produce a loud distorted/clipped sound.

When too high, the attenuation can start to decay before the max sample is output,
again causing clipping.

This PR keeps a queue of max sample values for the current buffered audio duration,
and avoids clipping if any sample in queue will clip.