Improve Human-Computer Interation of PVR EPG
With ShowEpgInfoOnEpgItemSelect set to false, pressing the "OK"
button over an EPG entry that is currently playing (i.e. Live),
the behaviour switches straight to that channel. However, EPG
entries that are not currently playing (i.e. future timeline entries)
then the "Programme Info" pop-up context menu appears to
give the user the option to set a timer for this future event.