XBMC-PVR: Add "Find similar programs" to Timers context menu
By adding the "Find similar programs" option, users are able to search
for upcoming showings of the same title.  This is particularly useful
for timers that are in "Conflict Error" with one another.  This provides
a quick method to re-shedule a conflicting timer.  The order of options
in the context menu has been re-ordered to reflect the most common actions
to resolve timer conflicts.