[PVR] Timeslot of EPG program
- In FixOverlappingEvents(), start time of EPG program from PVR client could be updated to fill the gap until
the previous program. It is a  problem for the PVR client when a new timer is created from EPG. Often the start
time of timeslot is use to identify the program. So here i update previous end time instead current start time
to fill the gap or close overlap. With these changes EPG program keep the same start time along the flow: from
the client, throw PVR and finally to the client when creating the timer.

- In GetTagAround() when previous program had no gap with current program (always true ? FixOverlappingEvents ?)
the previous was returned instead the current begins around. So here time should be strickly less.