Death to List Tyranny! Fix default viewtype selection.
Why in Heaven’s name do we need to falsify the List type for this to
work is beyond me. As long as it was <viewtype
label="List">list</viewtype> it hogged the default viewtype place,
persistently. Once changed to <viewtype label="List">panel</viewtype>
everything works as intended. PMS perhaps forcing a list view (when
found) as default?