Disable Channels main level List view, for now.
We can’t save the viewtype state in there as of today, so you can set
it to list and when you come back it’s panel view again. So thought
better to disable it altogether in that screen for now (neither being
ideal, but maybe this a smaller nuisance).