[pvr] Fix missing recordings
This fixes an issue with missing recordings that occurs under following
circumstances:

- Recording A in subdirectory "TV Show XY"
- Recording B in subdirectory "TV show xy"

Because of the difference in case, recording B won't show up in any
sub folder. This can happen due to EPG inconsistencies.

The presented sub folders are created with case-insensitive matching
using CFileItemList::Contains in GetSubDirectories.
But when determining if a recording belongs into a directory, the
comparison is case-sensitive (in IsDirectoryMember).

This commit changes the IsDirecroryMember check to be case-insensitive.