[musicdb] If we're rescanning an album from tags (i.e. tags have changed) make sure we update the album as best we can. This includes: 1. Update genre, year, compilation. 2. If a MBID exists, also update album name and album artist (as these may differ as we match on MBID). 3. Reset lastScraped time so that online metadata is refreshed. 4. Ensure we remove previously assigned album artists before adding the current ones. 5. Ensure we remove previously assigned album genres before adding the current ones.
The main place this process can fail is in multi-folder albums. In this case, the artists and genres (and year and compilation flag)
from the last scanned folder will be what is included.

A way to work around this for the future might be that, in the scanner, we not only scan the current folder, but for each album, we
grab all other folders those albums appear on and scan them at the same time.  This ensures that each album is complete during an
individual scan.