[players] Fix for subtitle flood at start of file
The test for pts==0 results in every subtitle in the overlay container being rendered at start of file.

This results in hundreds of subtitles being rendered which is slow and ugly. See:
https://dl.dropboxusercontent.com/u/3669512/temp/screenshot001.png

Remove this test.

Also remove setting the media_pts to the predicted media_pts as that causes a subtitle flurry after seeking while GPU fifo fills