Change the string comparison in CGUITextLayout::UpdateW to be case-sensitive
It seems likely that it was only case-insensitive due to the the default
seting of CStdStr::Equals. It's unlikely that you'd ever want to update a
string with another which is the same except for the case of one or more
characters, but if you did, I doubt you'd want to use the cached version if
so. And you're slowing down the comparison in all other cases to achieve that
dubious behaviour.