Add SmallStepForward to match SmallStepBack, bind to forward slash and add appropriate advanced setting.
It's handy to be able to skip a small amount forward too - e.g. if you have gone back a bit too far with step back default 30 secs).

I am not sure if it is ok to change all the key numbers in hey.h so have used a higher available value for now.