[pvr] fix: missing action mapping for ACTION_RECORD
This adds the missing button translator action "record" poiting to
ACTION_RECORD which is currently used in several PVR related windows,
but could not be used in key mappings.

It also adds a default implementation to GUIWindow which calls the
application player record functionality, so there is no need to use
XBMC.PlayerControls(record) in key mapping anymore.

Furthermore the default mapping for the record button in remote.xml is
changed to this new action.