[omxplayer] Always enable timestamp fifo mode.
hints.ptsinvalid is only set for h264 in avi files. However it also can affect any videos with B frames in any container
(we have reports of mp4 and mkv stuttering because of invalid pts values).

To solve this we always enable the timestamp fifo mode, and the gpu treats this as an optional flag, and falls back to non-fifo timestamp
mode if the variance of that timestamp stream is better.