Fix several issues with analytics data.
Fixes playback time tracking. We can’t rely on the offset stuff we need
track it manually. This make the value correct when we pause as well.

Send tracking event if a new item is played even if we didn’t see a
stop event.

Extracted the real item for Photos so we can get data.

Fixes #1023