Further performance improvements for updates.
Split artwork changes into altered and removed.

Add some log messages to allow for better timing.

In SetArtForItem only update an item of art if the url has actually
changed.  Saves a number of extra round trips.

Remove code that set the lastplayed and playcount to match any other
found copy of the movie.  It doesn't seem to make sense to do so during
an update, as the update requires the file already exists, so it's not a
new file.