Make URIUtils::IsInPath case insensitive
This fix mainly /vfs http handler to allow access to file in sources that where recreated with a different case.
(In this case an update / rescrape will not update the path if only the case changes)

This keep the behavior in sync with CUtil::GetMatchingSource and all other use case.