[Linux] add basic support for Wii U Pro Controller
Since Linux kernel 3.13, the Wii U Pro Controller works out-of-the-box
by simply connecting it as a bluetooth device. This is way better than
the Xbox 360 controller because you don't need a separate adapter.
Additionally the LEDs work properly and the controller has a power button,
so you can actually turn it off instead of having to unplug the battery.

This patch adds basic support for menu navigation and video playback. It
mostly matches the Xbox 360 Controller mappings (from the Wiki) with
some minor differences (e.g. the home button actually takes you home and
the secondary shoulder buttons are not analog).

I also tested this controller on Windows 8.1. You can connect it as a
bluetooth device but it doesn't actually work. The LEDs on the
controller keep blinking and XBMC only detects it as a joystick with 0
buttons, 0 hats and 0 axes. I guess there is simply no driver support but I
haven't looked into it in detail.