When constructing the photo transcoder URL we need to to check server port.
Before it was hardcoded to use 32400, we now check if we have a local
connection for the server and use that to lookup the port. We can’t
just use any activeConnection because it could be a port that is mapped
via UPnP and will not correspond to the port the server is actually
listening to.

Related to #1083.