Refactored PlayQueueManager.
This makes it possible to handle Local and Remote playQueues with the
same API and also show the PlayQueue on the HomeWindow at all times. It
now also restores the PlayQueue when the sessions ends.

Clicking the PlayQueue from the home window also takes it to
PlexPlayQueue.xml

Related to #1108
Related to #1106