[dxva] Close the DXVA2 decoder before opening again
This fix the issue described here: http://forum.xbmc.org/showthread.php?tid=190254

Basically if the current hardware context is dxva2, close it, as we're
about to open a fresh context.  I suspect something about ffmpeg is
causing a double call into GetFormat, which isn't present in frodo.

Perhaps a bug/change in behaviour with ffmpeg 1.2?