Revert the changes to primaryFilterActivated.
The primaryFilterActivated property will be set to “yes” when we don’t
have any secondaryFilters to show. In Movie Sections this means that we
have anything but the “All” primary filter marked. In Music sections
this means that we have Recently Added marked. Artists/Albums are still
able to be filtered.