Don't reload the playQueue if no changes occurred.
This makes two important changes to the playQueueUpdated function.

First one is to stop sending the PLAYQUEUE_UPDATED message if
reconcileplayqueue does not find any changes after the merge.

The second one is to actually use the playQueueItemId instead of key
when we do the merging. This is important because the same item can be
queued multiple times.

Test cases are updated accordingly.