Fix double back in music sections and by Album
At this point I am just doing hacks on hacks ;) The primary filter will
get back with album, but the real URL that we load is /all?type=9 so
they won’t match up, that leads to the double load on exit. Just work
around it for now. At some point I might redo the filter code to not be
that janky.