Limit the number of Genres to 2.
The problem we had in #1149 is that the video player hit the metadata
key and got ALL the genres and the listing above only includes two of
them. Which was a bit jarring. To get all genres in the listing I would
have to do a HTTP call per movie in the listing, which is not really
economical. So for now we limit it to two everywhere. Which kind of
makes sense.