Play everything from a PlayQueue
Create a playQueue even if we just play one item.
This had some far-reaching implications, especially for our old
PlexContentPlayerMixin that was not really going to work for PlayQueues.

The Resume/Select Media Part dialogs was moved to GUIDialogPlexMedia
and is called from MediaDecisionEngine to be able to show that when
it’s coming from a queue. This will only show for the first item in a
queue.

By removing PlexContentPlayerMixin a lot of classes had to be changed
because of include files and also because they had to call the
playQueueManager to play items instead.

This also lead to some refactoring of the way we play photos since
these are not supported by PlayQueues right now.

Fixes #1140