Add support for VideoPlayer.Offset
This works just a MusicPlayer.Offset. It involved some really nasty
gymnastics in info manager since the info manager function that
translated video item information was hardcoded to use m_currentFile. I
had to C&P that method and make it take an optional argument for
sending in a different FileItem.