CGUIDialogAudioSubtitleSettings: avoid not being able to change audio/subtitle streams
The cause of the issue is described in http://trac.xbmc.org/ticket/15210 with
additional discussion in https://github.com/xbmc/xbmc/pull/4734. This removes
the problematic logic for now which results in the same behaviour as we have
in Gotham (i.e. no worse no better) but this also means that we'll still show
outdated information for audio/subtitle streams if they are changed in the
background while the dialog is being shown.