Allow queueing from the ContextMenu when you don't have a PQ
The old logic only showed Play Next and Add to Up Next if we had a play
queue of the same type of item that you where trying to queue. So there
where no way to easily create a video play queue when you had played
music at some point.

This changes doe two things, it stops playback when we create a new
queue and it shows the context button for Play Next if you don’t have a
queue of the same type. The Add to up next is only shown if you
actually have a play queue in the correct type.

@mmccurdy07 sanity check on this behaviour would be appreciated.