Make sure that Win32DelayedDllload is compiled as a main object
This is a old bug fixed at some time, but when I went from building the
main binary directly to have a Object Library called PlexMain this was
not compiled correctly, maybe.

This is just a theory to why it could make some systems crash on start.
Let’s try it out :)

Related to #1201