[ADDONS][PVR] Avoid incorrect disabling PVR addons
When addons are installed from zip they aren't added to the addon
table in AddonsDB (Maybe this is the real issue here?).  When
PVRManager is started it checks only whether the AddonsDB has the addon,
which it doesn't in this case and the addon is marked disabled in
AddonsDB.  The addon and PVR both work fine until the next launch of
app, at which point the "No PVR Addon enabled!" dialog is thrown in the
user's face.  Marking the addon enabled here then adds it to the db as
expected.