Fix preplay of cloud synced movies
This was because of the stupid way we detected single items. Before we
where trying to figure it out via the URL, but that was cumbersome and
not very nice. Now we instead check if the container have a size of 1,
if it does we also check if the viewGroup property is set and in that
case it’s “always” a container and not a single item.