Introducing Plex TextureCache
This is an Plex Implementation of TextureCaching. It inherits from the XBMC implementation which uses a Database. Plex doesn't need to have a database and some benchmarks have proven that using the database makes everything crawl.

We therefore only use the Thumbnails local storage which is enough for our use.