Fix advancing video playQueues.
There where a few fishy things here.

First the debug log was totally wrong and lied to me which
made me a bit confused. That was because I was using the wrong
variable in to show what playlist was in use.

Then I added the playQueueItemID to the timelines going to the
server as well. This makes the code on the server lookup more
reliable.

Lastly the media item that we got from the MDE had the wrong itemId
because it was using ratingKey instead of playQueueItemID.

Fixes #1260