[fanart] use FirstChild()->Value() rather than unchecked GetText().
NOTE: Potential change in behaviour in that if a url attribute is specified in the <fanart>
node and <thumb> children are specified without a value, it used to just take url as the
url.  This was not intended, however, as url is supposed to be the base url, not the full url.