Fix Cannot jump to home from Extras strip #1283
When you have a scrolling list, Application will try first to send the message to the window rather than processing it.
in the case of 'h' key, and on preplay window, the BUILTIN message will come down to preplay window and be handled by a controlgroup that will put it in an onclick message forwarded to the controls. this is pretty much insane.

So we just say that we dont wanna handle builtins, then Application will handle them properly.