Add PQ EditMode To PQ Window.
EditMode is only available when playlist is not playing and can be entered with
- Using long Ok press will enter / exit EditMode
- Context menu and `Toggle Edit Mode`

Once in EditMode, arrows will allow to move items around.

`ACTION_MOVE_ITEM_UP` and `ACTION_MOVE_ITEM_DOWN` are now handled and will move up & down the selected item in the PQ without entering EditMode.