Changes to PlexConnection::Equals and ::Merge
The use case for the two changes below are the following: We get a local connection to a server from plex.tv before we get a local discovery connection with GDM. That would lead to two connections to the same URL and more connection testing that is not needed.

To avoid that we make two changes in PlexConnection:

1) It changes the way two connections equals when the only thing that separates them is the token. It allows the overall match to be true if one of the connections has a empty token. This allows the code to try to merge a tokenless connection with one with a token.

2) It changes the merge function to not overwrite a token with a empty one. Since we prefer tokens.

Unit tests are now testing all around this code as well.

Fixes #1310