fixed: keyboard layout blank and double QWERTY.
There are 3 solutions to this issue

1) @notspiff fix to use a fallback to English with the font13 which is
not what's used elsewhere in skin
https://github.com/notspiff/xbmc-cmake/commit/8ca6d81189acb43d1d136935b5013b5f9ae5410c

2) use the Roboto fonts I was working on and no longer am, but already has
the Hebrew characters added (correctly or incorrectly idk)
http://forum.xbmc.org/showthread.php?tid=150067

3) simple, use English names, since default language is English, users
must have some degree of understanding of that and smart enough to know it.

Since I no longer work on that font , nor speak such languages I was
adding characters for and have no desire to extend it from what is done
atm

This was my solution, both simplest and avoids dealing with niggling issues.