Multiples Playqueues : replace PQ interface with base class, part of #1214
This patch replaces the IPlexPlayQueueBase with a CPlexPlayQueue, in order to ba able to hold a PQ map based on type in the PQ manager.

It moves the type and PQ version to the CPlexPlayQueueVersion rather than keeping it in PQ manager.

It also changes a bunch of the PQmanager API so that functions take a PQ type as additionnal parameter.