[confluence] fixed changing theme breaks background
This fixes trac.xbmc.org/ticket/15429 which I opened myself.

The root cause is not fully understood and is probably not fixed.

However with this change the expected behaviour is normal and changing
themes whatever the theme name is no longer breaks the totally unrelated
background with this fix.

If someone feels like investigating and making a proper fix please go
ahead as I dont understand the underlying C++ and skinning is really not
my speciality.

@ronie