[PiSink] Dual audio support
The Pi supports dual audio in omxplayer, where the gpu outputs audio to
both hdmi and analogue connections.

With dvdplayer being more usable on Pi now, there are frequent requests
for this functionality through the Pi Sink.

This removes the gui setting for "dual audio" and creates a new sink
output for dual audio. It add support to the Pi sink and makes omxplayer
use the new setting.