[mmalcodec] Introduce a preroll period to buffer up frames on startup
There is an issue with current player/render system where everything starts from empty.
This means stuttering and frame drops are more likely at the start of a file as you
try to fill up the queue of decoded pictures given to renderer.

To mitigate this somewhat, we avoid returning pictures from mmal codec until we
have a few queued up.