[AE] Allow resampler to indicate when it has too much data buffered
The Pi resampler doesn't want more samples submitted when it has a whole input buffer queued,
and drops data in this case. This can be triggered when video clock is enabled and audio is resampled.

The FFMPEG resampler buffers more data internally and wants up to two input buffers submitted.

Add a method to query if there is space for more samples, and return approprate values.