[android] Add 7.1 PCM playback support
Add Android 7.1 PCM audio playback support. This support requires
Android 5.0 device.

When support is not available, the code falls back to 5.1 and stereo
playback.

Tested on ADT-1.

TODO: Channel layout is currently messed up on at least ADT-1:
https://code.google.com/p/android-developer-preview/issues/detail?id=1800

What actions are needed to handle that (if any) depends on whether it
will be fixed for 5.0 final or if this affects other hardware.

The code still restricts channel layout selection to 2.0/5.1/7.1 since
all other layouts seem to still be downmixed to stereo.